% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{source_data}
\alias{source_data}
\title{Funciton to source Covid-19 India datasets from the COVID-19 India API}
\usage{
source_data()
}
\value{
this function returns a list of data-sets sourced from
\url{https://api.covid19india.org/csv/latest/}
}
\description{
Funciton to source Covid-19 India datasets from the COVID-19 India API
}
\note{
This function does not require any parameters to be passed.
}
